Fixes #2 Adds ability to alias fields
Adds support for the ability to give additional names to fields in a
struct. This includes adding a dependency on the master branch of the
attribution crate. Attribution is to support more complex forms of the
helper attribute. Later when a new version of attribution is released,
the dependency should be updated to reference the version hosted on
crates-io.