Updates the `HasField` trait and macro
Updates `HasField` to separate the trait into two traits, `BaseHasField`
and `HasField`. The former is the trait that is actually derived with
provides the bare essential information for all other parts of the
trait which are useful to the library consumer.

This bare essential information is the type of the field that the struct
contains, and a way in which to calculate the offset of the field
measured in bytes.

The second trait (`HasField`) contains all parts of the interface that a
library consumer might actually want to invoke.